executor: merge plugin shim with executor package
