+++ drivers/shared/executor/utils_windows.go
@@ -1,0 +1,6 @@
+package executor
+
+import "os/exec"
+
+// TODO Figure out if this is needed in Windows
+func isolateCommand(cmd *exec.Cmd) {}
